+++ lib/src/controls/flyouts/menu_flyout.dart
@@ -320,6 +320,7 @@ class MenuFlyoutItem extends MenuFlyoutItemBase {
     return FlyoutListTile(
       margin: EdgeInsets.zero,
       selected: selected,
+      showSelectedIndicator: false,
       icon: leading ??
           () {
             if (_useIconPlaceholder) return const Icon(null);